introducing interface for window management
